marketplace/requirements.txt
flask~=3.0.0
grpcio-tools~=1.30
2to3~=1.0.0
Jinja2~=3.1.2
pytest~=7.4